Project Update

+++ app/Http/Controllers/backend/TecnicianController.php
@@ -38,7 +38,7 @@ class TecnicianController extends Controller
         $tecnician->Type_of_work = $request->Type_of_work;
 
         if (isset($request->image)) {
-            $imageName = rand().'.'.'tecnician'.'.'.$request->image->extension();
+            $imageName = rand() . '.' . 'tecnician' . '.' . $request->image->extension();
             $request->image->move('backend/images/tecnician/', $imageName);
             $tecnician->image = $imageName;
         }
@@ -59,15 +59,50 @@ class TecnicianController extends Controller
         return view('backend.tecnician.view', compact('tecnician'));
     }
 
-    public function tecnicianDelete($id) 
+    public function tecnicianDelete($id)
     {
         $tecnician = Technician::find($id);
-        
-         if($tecnician->image && file_exists('backend/images/tecnician/'.$tecnician->image)){
+
+        if ($tecnician->image && file_exists('backend/images/tecnician/' . $tecnician->image)) {
+            unlink('backend/images/tecnician/' . $tecnician->image);
+        }
+
+        $tecnician->delete();
+        return redirect() - back();
+    }
+
+    public function tecnicianEdit($id)
+    {
+        $tecnician = Technician::find($id);
+        return view('backend.tecnician.edit',  compact('tecnician'));
+    }
+
+    public function tecnicianUpdate(Request $request, $id)
+    {
+        $tecnician = Technician::find($id);
+
+        $tecnician->name = $request->name;
+        $tecnician->phone = $request->phone;
+        $tecnician->dob = $request->dob;
+        $tecnician->nid = $request->nid;
+        $tecnician->passport_no = $request->passport_no;
+        $tecnician->present_address = $request->present_address;
+        $tecnician->permanent_address = $request->permanent_address;
+        $tecnician->join_date = $request->join_date;
+        $tecnician->Type_of_work = $request->Type_of_work;
+        if (isset($request->image)) {
+
+            if ($tecnician->image && file_exists('backend/images/tecnician/' . $tecnician->image)) {
                 unlink('backend/images/tecnician/'.$tecnician->image);
             }
-        
-        $tecnician->delete();
-        return redirect()-back();
+
+            $imageName = rand() . 'Technician'.'.'.$request->image->extension();
+            $request->image->move('backend/images/tecnician/', $imageName);
+
+            $tecnician->image = $imageName;
+        }
+
+        $tecnician->save();
+        return redirect()->back();
     }
 }
